+++ src/ga4/Analytics.php
@@ -93,6 +93,14 @@ class Analytics
      */
     public function addEvent(AbstractEvent $event): BaseRequest
     {
+        $clientId = $this->request()->getClientId();
+
+        if (strpos($clientId, '.') !== false) {
+            [$sessionId, $sessionNumber] = explode('.', $clientId);
+            $event->setSessionId($sessionId);
+            $event->setSessionNumber($sessionNumber);
+        }
+
         return $this->request()->addEvent($event);
     }