+++ src/models/Settings.php
@@ -70,6 +70,11 @@ class Settings extends Model
     public $createGclidCookie = true;
 
     /**
+     * @var int The session duration as set in GA4 admin settings.
+     */
+    public $sessionDuration = 30;
+
+    /**
      * The field in a Commerce Product Variant that should be used for the
      * category
      *